Rename class from backups to LocalBackupsTasks
