+++ GithubBackup/Class/ApplicationStatus.cs
@@ -81,8 +81,8 @@ namespace GithubBackup.Class
                     Globals._mailfrom,
                     Globals._mailto,
                     emailStatus,
-                    Globals.repocountelements,
-                    Globals.repoitemscountelements,
+                    Globals._repocountelements,
+                    Globals._repoitemscountelements,
                     Globals._repoCount,
                     0,
                     Globals._backupFolderName,
